views 에 post일때만 작동하도록 if 문 추가

+++ django/blog/views.py
@@ -96,8 +96,9 @@ def post_delete(request, pk):
     :return:
     """
     # pk에 해당하는 post를 삭제
-    post = Post.objects.get(pk=pk)
-    post.delete()
+    if request.method == 'POST':
+        post = Post.objects.get(pk=pk)
+        post.delete()
 
-    # 이후 post-list 라는 URL name 을 갖는 view로 redirect
-    return redirect('post-list')
+        # 이후 post-list 라는 URL name 을 갖는 view로 redirect
+        return redirect('post-list')
